a big turn of events in the history of build_install_s
NEW: verbose is now private

UPDATE: there's no more hell.builds and hell.installs. It's all stored
in _G now

NEW: we get builds_installs from the target itself (the table). As this
is, if we define a target in an inner script with a new _ENV, it doesn't
work (TODO: work this out, maybe a __newindex?)

NEW: we can now susbstitute any string, for Builders (hellutils.lua:subst)