a little something in dependency, hell.outdir, prefix input paths
NEW: input paths are prefixed in the build, so you write the relative
path in the script, and hell uses the path relative to the PWD.
  TODO: prefix input using shell rules for arguments, as opposed to
  prefix everything after a blank [ \t\n]

NEW: output are prefixed with the output directory for shadow builds,
and the keepDirStructure is already handled
  TODO: test some more

NEW: dependencies! It's tested inside the Lua part, but will rather be
handled by Haskell.
  TODO: test aswell, might not be working properly