-c|--ignore-cyclic-dependency; -n doesn't `checkNeedRebuild ()`
NEW: finding a cyclic dependency now aborts building. Maybe one needs
that to be ignored, so added a new flag for that (it's got a nice
explanation at `hell -H` xP)

UPDATE: when running a dry run (-n|--no-exec), suppose all builds should
be run, so we check all of the commands.